Inline AuditTrailChange into AuditTrailItemComponent
- Remove last_change from AuditTrailChange, as CSS can do the styling
  for us
- AuditTrailItem#changes now returns a list of AuditTrailChanges
- Old markup in audit_trail_change_component now taken care of by
  govuk_summary_list
