--- app/components/support_interface/audit_trail_change_component.html.erb
@@ -1,8 +1,0 @@
-<div class="govuk-summary-list__row">
-  <dt class="govuk-summary-list__key" style="<%= style %>">
-    <%= attribute %>
-  </dt>
-  <dd class="govuk-summary-list__value" style="<%= style %>">
-    <%= format_audit_values %>
-  </dd>
-</div>
